pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metareason"
version = "0.1.0"
description = "A meta-reasoning framework for LLM sampling and oracle evaluation"
authors = [
    {name = "Jeff Bradley", email = "admin@metareason.ai"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
keywords = ["llm", "meta-reasoning", "sampling", "evaluation", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    # Core dependencies with proper version pinning
    "pymc>=5.0,<6.0",  # Bayesian analysis
    "arviz>=0.18,<1.0",  # Bayesian diagnostics and visualization
    "numpy>=1.24,<2.0",  # Numerical computing
    "scipy>=1.11,<2.0",  # Scientific computing
    "pydantic>=2.0,<3.0",  # Configuration validation
    "pydantic-settings>=2.0,<3.0",  # Settings management
    "click>=8.1,<9.0",  # CLI framework
    "pyyaml>=6.0,<7.0",  # Configuration parsing
    "jinja2>=3.1,<4.0",  # Template rendering
    "httpx>=0.25,<1.0",  # Async HTTP client
    "matplotlib>=3.8,<4.0",  # Visualization
    "seaborn>=0.13,<1.0",  # Statistical visualization
    # Additional useful dependencies
    "rich>=13.0,<14.0",  # Rich terminal output
    "pandas>=2.0,<3.0",  # Data manipulation
    "tqdm>=4.66,<5.0",  # Progress bars
    "aiohttp>=3.9,<4.0",  # Async HTTP server/client
    "ollama>=0.5,<1.0",  # Ollama API client
    "openai>=1.0,<2.0",  # Official OpenAI API client
    "anthropic>=0.34,<1.0",  # Official Anthropic API client
    "google-genai>=1.0,<2.0",  # Official Google GenAI SDK for Gemini
    "tenacity>=8.2,<9.0",  # Retry logic
    "structlog>=24.0,<25.0",  # Structured logging
    "python-dotenv>=1.0,<2.0",  # Environment variable management
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-asyncio>=0.21,<1.0",
    "pytest-cov>=4.1,<5.0",
    "pytest-mock>=3.12,<4.0",
    "pytest-xdist>=3.5,<4.0",
    "black>=24.0,<25.0",
    "isort>=5.13,<6.0",
    "flake8>=7.0,<8.0",
    "flake8-docstrings>=1.7,<2.0",
    "flake8-bugbear>=24.0,<25.0",
    "bandit>=1.7,<2.0",
    "pre-commit>=3.6,<4.0",
    "ipython>=8.18,<9.0",
    "jupyter>=1.0,<2.0",
    "notebook>=7.0,<8.0",
]
docs = [
    "sphinx>=7.2,<8.0",
    "sphinx-rtd-theme>=2.0,<3.0",
    "sphinx-autodoc-typehints>=1.25,<2.0",
    "myst-parser>=2.0,<3.0",
]

[project.scripts]
metareason = "metareason.cli.main:metareason"

[project.urls]
Homepage = "https://github.com/jeffgbradley2/metareason-ai"
Repository = "https://github.com/jeffgbradley2/metareason-ai/metareason-core"
Documentation = "https://metareason.readthedocs.io"

[tool.hatch.version]
path = "src/metareason/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/metareason"]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
src_paths = ["src", "tests"]


[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=metareason",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["src/metareason"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]
